app/requirements.txt
APScheduler==3.9.1
databases==0.6.2
dependency-injector==4.40.0
greenlet==2.0.1
psycopg-binary==3.1.4
pyaes==1.6.1
pyasn1==0.4.8
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
rsa==4.9
six==1.16.0
SQLAlchemy==1.4.41
Telethon==1.25.4
tzdata==2022.6
tzlocal==4.2
asyncpg==0.27.0
psycopg2-binary==2.9.5
schedule==1.1.0
PySocks==1.7.1